streamlit/requirements.txt
streamlit
psycopg2
scipy
#logging
psycopg2
numpy
gymnasium
matplotlib
scikit-learn
shap
panda
lime